bots/requirements.txt

modal>=0.55.0
fastapi>=0.115.0
uvicorn>=0.20.0
google-cloud-firestore>=2.19.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
httpx>=0.28.0
pydantic>=2.10.0
openai>=1.50.0
python-dotenv>=1.0.0
